requirements.txt
pdfplumber
PyMuPDF
pdf2image
pytesseract
regex
python-dateutil
pycountry
rapidfuzz
pytest
